Update to fix broken phone build. Added build solution to build all related projects and tests (does not include examples). This build should work for both desktop Silverlight and Phone Silverlight.

+++ src/SterlingSln/Wintellect.Sterling.IsolatedStorage/IsoStorageHelper.cs
@@ -36,7 +36,12 @@ namespace Wintellect.Sterling.IsolatedStorage
         {
             if (_iso != null) return;
 
-            _iso = siteWide ? IsolatedStorageFile.GetUserStoreForSite() : IsolatedStorageFile.GetUserStoreForApplication();
+            _iso = 
+#if WINDOWS_PHONE
+#else
+                siteWide ? IsolatedStorageFile.GetUserStoreForSite() :
+#endif
+                IsolatedStorageFile.GetUserStoreForApplication();
         }
 
         /// <summary>
@@ -199,7 +204,7 @@ namespace Wintellect.Sterling.IsolatedStorage
                 }
 
                 var dirPath = path.TrimEnd('\\', '/');
-                if (_iso.DirectoryExists(dirPath))
+                if (!string.IsNullOrEmpty(dirPath) && _iso.DirectoryExists(dirPath))
                 {
                     _iso.DeleteDirectory(dirPath);                   
                 }
